fix up typings and usages, etc
