+++ packages/common/tsconfig.json
@@ -3,7 +3,8 @@
 	"compilerOptions": {
 		"composite": true,
 		"outDir": "./dist/esm",
-		"rootDir": "./src"
+		"rootDir": "./src",
+		"lib": ["esnext", "dom"]
 	},
 	"include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.mts", "./typings"]
 }
